CTCP-5582: Refactor based on Jack's comments
