+++ app/utils/IncidentP5Helper.scala
@@ -31,10 +31,30 @@ class IncidentP5Helper(
   data: IncidentType03,
   refDataService: ReferenceDataService
 )(implicit messages: Messages, ec: ExecutionContext, hc: HeaderCarrier)
-    extends IncidentP5MessageHelper(refDataService)
+    extends DeparturesP5MessageHelper
     with Logging {
 
   private val displayIndex = data.sequenceNumber
+
+  def incidentCodeRow: Future[Option[SummaryListRow]] =
+    refDataService.getIncidentCode(data.code).map {
+      incidentCode =>
+        buildRowFromAnswer[String](
+          answer = Some(incidentCode.toString),
+          formatAnswer = formatAsText,
+          prefix = "departure.notification.incident.index.code",
+          id = None,
+          call = None
+        )
+    }
+
+  def incidentDescriptionRow: Option[SummaryListRow] = buildRowFromAnswer[String](
+    answer = Some(data.text),
+    formatAnswer = formatAsText,
+    prefix = "departure.notification.incident.index.description",
+    id = None,
+    call = None
+  )
 
   def countryRow: Future[Option[SummaryListRow]] =
     refDataService.getCountry(data.Location.country) map {
@@ -88,12 +108,12 @@ class IncidentP5Helper(
   def incidentInformationSection: Future[StaticSection] =
     for {
       countryRowOption <- countryRow
-      incidentCodeRow  <- incidentCodeRow(data.code, "departure.notification.incident.index.code")
+      incidentCodeRow  <- incidentCodeRow
     } yield StaticSection(
       sectionTitle = None,
       rows = Seq(
         incidentCodeRow,
-        incidentDescriptionRow(data.text, "departure.notification.incident.index.description"),
+        incidentDescriptionRow,
         countryRowOption,
         identifierTypeRow,
         coordinatesRow,
