+++ test/helper/IncidentP5HelperSpec.scala
@@ -23,6 +23,7 @@ import models.{Country, IncidentCode, RichAddressType18}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalacheck.Arbitrary.arbitrary
+import org.scalacheck.Gen
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 import play.api.inject
 import play.api.inject.guice.GuiceApplicationBuilder
@@ -48,6 +49,48 @@ class IncidentP5HelperSpec extends SpecBase with ScalaCheckPropertyChecks with G
     val incidentType03 = arbitraryIncidentType03.arbitrary.sample.value
 
     "rows" - {
+
+      "incidentCodeRow" - {
+        "must return a row" in {
+          forAll(Gen.alphaNumStr) {
+            value =>
+              val incidentCode = IncidentCode(
+                "1",
+                "The carrier is obliged to deviate from the itinerary prescribed in accordance with Article 298 of UCC/IA Regulation due to circumstances beyond his control."
+              )
+
+              val incidentType = incidentType03.copy(code = value)
+
+              when(refDataService.getIncidentCode(any())(any(), any()))
+                .thenReturn(Future.successful(incidentCode))
+
+              val helper = new IncidentP5Helper(incidentType, refDataService)
+              val result = helper.incidentCodeRow.futureValue.value
+
+              result.key.value mustBe "Incident code"
+              result.value.value mustBe
+                "1 - The carrier is obliged to deviate from the itinerary prescribed in accordance with Article 298 of UCC/IA Regulation due to circumstances beyond his control."
+              result.actions must not be defined
+          }
+        }
+      }
+
+      "incidentDescriptionRow" - {
+        "must return a row" in {
+          forAll(Gen.alphaNumStr) {
+            value =>
+              val incidentType = incidentType03.copy(text = value)
+
+              val helper = new IncidentP5Helper(incidentType, refDataService)
+              val result = helper.incidentDescriptionRow.value
+
+              result.key.value mustBe "Description"
+              result.value.value mustBe value
+              result.actions must not be defined
+          }
+        }
+      }
+
       "countryRow" - {
         "must return a row with description when ref data look up is successful" in {
           when(refDataService.getCountry(any())(any(), any()))
